Reservations
The user is now able to reserve a running movie. Their reservation will also be saved in a json file. The reservations are sorted by users emailadress. I  also created a factory file which creates new instances of logics, so that you dont have to create a new one everytime you use a file